requirements.txt
Cython==0.29.23
h5py==2.10.0
nltk==3.6.1
nose==1.3.7
numpy==1.20.1
pandas==1.2.4
scikit-learn==0.24.1
scikit-optimize==0.8.1
scipy==1.6.2
tqdm==4.59.0
lightfm==1.16
dwave-ocean-sdk==3.1.1